+++ admin/views/themes/tastyigniter-blue/mail_templates_variables.php
@@ -7,7 +7,8 @@
 			</div>
 			<div class="modal-body">
 				<?php foreach ($variables as $key => $value) { ?>
-					<div class="panel">
+					<?php $filter_class = !empty($filters[$key]) ? implode($filters[$key], ' ') : ''; ?>
+					<div class="panel hide <?php echo $filter_class; ?>">
 						<div class="panel-heading border-bottom"><h4 class="panel-title"><?php echo $key; ?></h4></div>
 						<div class="list-group supported-var">
 							<?php foreach ($value as $variable) { ?>
@@ -23,3 +24,17 @@
 		</div>
 	</div>
 </div>
+<script type="text/javascript">
+	$(document).ready(function() {
+		$('#mail-variables').on('show.bs.modal', function (event) {
+			var button = $(event.relatedTarget);
+			var filter = button.data('filter');
+			// If necessary, you could initiate an AJAX request here (and then do the updating in a callback).
+			// Update the modal's content. We'll use jQuery here, but you could use a data binding library or other methods instead.
+
+			var modal = $(this);
+			modal.find('.modal-body > .panel').addClass('hide');
+			modal.find('.modal-body > .panel.' + filter).removeClass('hide')
+		})
+	});
+</script>